Move players to new package players and bots to a subpackage within players named bots * Move Engine class to its own engine package * Remove unecessary/empty BotPlayerTest and PieceTest classes and their respective packages
