+++ src/main/java/com/abdmoh123/chessgame/players/Player.java
@@ -1,6 +1,6 @@
-package com.abdmoh123.chessgame.control;
+package com.abdmoh123.chessgame.players;
 
-import com.abdmoh123.chessgame.control.engine.Engine;
+import com.abdmoh123.chessgame.engine.Engine;
 import com.abdmoh123.chessgame.moves.Move;
 
 public abstract class Player {
