+++ src/main/java/com/abdmoh123/chessgame/players/bots/BotPlayer.java
@@ -1,4 +1,6 @@
-package com.abdmoh123.chessgame.control;
+package com.abdmoh123.chessgame.players.bots;
+
+import com.abdmoh123.chessgame.players.Player;
 
 public abstract class BotPlayer extends Player {
     protected BotPlayer(boolean is_white_in) {
