+++ src/main/java/com/abdmoh123/chessgame/players/bots/RandomBot.java
@@ -1,7 +1,7 @@
-package com.abdmoh123.chessgame.control;
+package com.abdmoh123.chessgame.players.bots;
 
 import com.abdmoh123.chessgame.boards.Space;
-import com.abdmoh123.chessgame.control.engine.Engine;
+import com.abdmoh123.chessgame.engine.Engine;
 import com.abdmoh123.chessgame.moves.Move;
 
 import java.util.ArrayList;
