+++ src/test/java/com/abdmoh123/chessgame/engine/EngineTest.java
@@ -1,4 +1,4 @@
-package com.abdmoh123.chessgame.control.engine;
+package com.abdmoh123.chessgame.engine;
 
 import org.junit.Before;
 import org.junit.Test;
