--- src/test/java/com/abdmoh123/chessgame/pieces/PieceTest.java
@@ -1,9 +1,0 @@
-package com.abdmoh123.chessgame.pieces;
-
-import org.junit.Before;
-
-public class PieceTest {
-    @Before
-    public void init() {
-    }
-}
